requirements.txt
spotipy
matplotlib
python-dotenv